ner-178,179,189,190-192 Change PW & Email and paper updates
