+++ src/screens/InstructorScreens/SettingsStackNavigator/SettingStackNavigatorInstructor.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import InstructorSetting from './InstructorSetting'
-import ChangePassword from './ChangePassword'
-import ChangeEmail from './changeEmail'
+import ChangePassword from '../../changeAuth/ChangePassword'
+import ChangeEmail from '../../changeAuth/ChangeEmail'
 
 const Stack = createNativeStackNavigator()
