+++ src/screens/SignUpScreen.js
@@ -7,15 +7,14 @@ const API_URL =  'http://localhost:3001/api/user/create';
 //import APIConnection from "../../../utility/APIConnection";
 
 export default function SignUpScreen({navigation}) {
-
+  const mailFormat = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
 
 
   const [selectedIndex, setSelectedIndex] = useState();
   const [check1, setCheck1] = useState(false);
+  const [isSecureEntry, setSecureEntry] = useState(true);
+  const [isSecureEntry2, setSecureEntry2] = useState(true);
 
-
-  //pasword eye
-  const [passwordEye, setPasswordEye] = useState(false);
 
 
   //TODO: Add proper set state on text input in order to recieve and update as user types their info
@@ -24,14 +23,23 @@ export default function SignUpScreen({navigation}) {
   const [user_password, setPassword] = useState('');
   const [user_type, setUserType] = useState('string');
   const [message, setMessage] = useState('');
-  
+  const [confirmPW, setConfirmPW] = useState(''); 
 
 
+  const emailError = () => {
+    if(user_email.match(mailFormat)){
+      return false; 
+    }else{
+      return true; 
+    }
+  }
+
+  const passwordHelper = () => {
+    return (user_password.length < 8); 
+  }
 
 
   const onSubmitHandler = () => {
-    console.log("email and pass and name: ",user_email, user_password, user_name, user_type);
-    
     if(user_name === null){
       alert("Name cannot be empty");
       return; 
@@ -45,9 +53,11 @@ export default function SignUpScreen({navigation}) {
       alert("Password needs to be at least 8 characters");
       return;
     }
-      
 
-
+    if(!user_password.match(confirmPW)){
+      alert("Passwords do not match.");
+      return; 
+    }
 
     const payload = {
         user_name,
@@ -82,13 +92,7 @@ export default function SignUpScreen({navigation}) {
 };
 
 
-  const emailError = () => {
-    return !user_email.includes('@');
-  }
 
-  const passwordError = () => {
-    return (user_password.length < 8); 
-  }
 
 
 
@@ -113,7 +117,7 @@ export default function SignUpScreen({navigation}) {
             outlineColor={'#4970FA'}
             left={<TextInput.Icon name='account'/>}
             onChangeText={newText => setName(newText)}
-          />
+        />
 
         <TextInput 
             style={styles.input}
@@ -128,20 +132,37 @@ export default function SignUpScreen({navigation}) {
           </HelperText>
 
         <TextInput 
-          secureTextEntry={true} 
-          type={passwordEye === false ? "password" : "text"}
+          secureTextEntry={isSecureEntry} 
           style={styles.input}
           label={'Password'} 
           mode={'outlined'}
           left={<TextInput.Icon name='lock'/>}
-          right={<TextInput.Icon name="eye"/>}
+          right={  
+              <TextInput.Icon name="eye"
+                onPress={() => setSecureEntry(prev => !prev)}
+              />
+          }
           outlineColor={'#4970FA'}
           onChangeText={newText => setPassword(newText)}
         />
-        <HelperText type='info' visible={passwordError()} style={{color:'#000'}}>
+        <HelperText type='info' visible={passwordHelper()} style={{color:'#000'}}>
             Password needs to be at least 8 characters long.
           </HelperText>
-        
+
+
+          <TextInput 
+          secureTextEntry={isSecureEntry2} 
+          style={styles.input}
+          label={'Confirm Password'} 
+          mode={'outlined'}
+          left={<TextInput.Icon name='lock'/>}
+          right={ <TextInput.Icon 
+            name="eye"
+            onPress={() => setSecureEntry2(prev => !prev)}
+          />}
+          outlineColor={'#4970FA'}
+          onChangeText={newText => setConfirmPW(newText)}
+        />
     
 
 
@@ -169,11 +190,10 @@ export default function SignUpScreen({navigation}) {
             color={'#4970FA'}
             checked={check1}
             onPress={() => setCheck1(!check1)}
-            style={styles.check}
             checkedColor={'#4970FA'}
             uncheckedColor={'#000'}
           />
-          <Text style={styles.boxLabel}>By checking, you Agree to our Terms of Use and understand our privacy policy. 
+          <Text>By checking, you Agree to our Terms of Use and understand our privacy policy. 
               You may recieve notifications via email.</Text>
         </View>
 
@@ -194,78 +214,9 @@ export default function SignUpScreen({navigation}) {
         
         
 
-      </View>
-      {/*
-      <KeyboardAvoidingView behavior={Platform.OS === "ios" ? "padding" : "height"} style={styles.form}> 
-
-        <TextInput 
-          style={styles.input}
-          label={'Name'} 
-          mode={'outlined'}
-          onChangeText={newText => setName(newText)}
-        />
-
-        <TextInput 
-          style={form.input}
-          label={'Email'} 
-          mode={'outlined'}
-          onChangeText={newText => setEmail(newText)}
-        />
-      
-        
-        <TextInput 
-          secureTextEntry={true} 
-          type={passwordEye === false ? "password" : "text"}
-          style={form.input}
-          label={'Password'} 
-          mode={'outlined'}
-          onChangeText={newText => setPassword(newText)}
-        />
-
-
-        <TextInput 
-          secureTextEntry={true} 
-          style={form.input} 
-          mode={'outlined'}
-          label={'Confirm Password'}
-        />
-      </KeyboardAvoidingView>
-          
-
-     <View style={button.container}>
-      <Text style={styles.status}>Are you a:</Text>
-        <ButtonGroup
-          buttons={['Student', 'Educator']}
-          selectedIndex={selectedIndex}
-          onPress={(value) => {
-            setSelectedIndex(value);
-            if(value == 0){
-              setUserType('student');
-            } else if(value == 1 ){
-              setUserType('instructor');
-            }
-          }}
-          containerStyle={button.selector}
-      />
-     </View>
-
-      <View style={styles.checkboxContainer}>
-        <CheckBox
-          center
-          checked={check1}
-          onPress={() => setCheck1(!check1)}
-          style={styles.checkbox}
-        />
-        <Text style={styles.boxLabel}>By checking, you Agree to our Terms of Use and understand our privacy policy. You may recieve notifications via email.</Text>
-      </View>
-
-      <TouchableOpacity disabled={!check1}  onPress={onSubmitHandler}>
-          <View style={register.registerButton}>
-            <Text style ={register.registerText}>Sign Up</Text>
-          </View>
-      </TouchableOpacity>
-        */ }
-    </View>
+        </View>
+       
+  </View>
 
 
 
@@ -307,16 +258,6 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     marginTop: 10
   },
-
-  check: {
-
-  },
-
-  boxLabel: {
-    
-  },
-
-  
 
 })
