+++ src/screens/StudentScreens/SettingsStackNavigator/SettingStackNavigatorStudent.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
-import StudentSetting from './StudentSetting'
-import ChangePassword from './ChangePassword'
-import ChangeEmail from './changeEmail'
+import StudentSetting from './StudentSetting';
+import ChangePassword from '../../changeAuth/ChangePassword';
+import ChangeEmail from '../../changeAuth/ChangeEmail';
 
 const Stack = createNativeStackNavigator()
