+++ src/screens/UserProfile/icons/defaultAvatar.png
